+++ statistics.h
@@ -2,8 +2,8 @@
 #define PAGE_REPLACEMENT_SIMULATOR_STATISTICS_H
 
 typedef struct statistics {
-    double average_memory_utilization_count_total;
-    double average_runnable_processes_count_total;
+    long average_memory_utilization_count_total;
+    long average_runnable_processes_count_total;
     int total_memory_references;
     int total_page_faults;
     long run_time;
@@ -12,11 +12,11 @@ typedef struct statistics {
 
 statistics *CreateStatistics(int max_page_size);
 
-void UpdateAverageMemoryUtilization(statistics *stats, int occupied_pf);
+void UpdateAverageMemoryUtilization(statistics *stats, long occupied_pf);
 
 void UpdateAverageRunnableProcesses(statistics *stats, int non_blocked);
 
-void UpdateTotalMemoryReferences(statistics *stats, int references);
+void UpdateTotalMemoryReferences(statistics *stats, long references);
 
 void UpdateTotalPageFaults(statistics *stats);
